% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backtrackLCS}
\alias{backtrackLCS}
\title{Retrieving from a matrix Longest Common Subsequence (LCS) between a pair of numeric vector.}
\usage{
backtrackLCS(x, y)
}
\arguments{
\item{x}{an integer vector}

\item{y}{an integer vector}
}
\value{
an integer with the length of Longest Common Subsequence (LCS)
}
\description{
This function retrieves the Longest Common Subsequence (LCS)
between two numeric vectors by backtracking the matrix obtained with dynamic programming.
}
\examples{
backtrackLCS( c(1,2,3,4,5),c(1,2,4))

}
