% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\docType{class}
\name{BCUnibic-class}
\alias{BCUnibic-class}
\title{Class BCUnibic}
\description{
An S4 class to represent \code{\link{BCUnibic-class}} UniBic biclustering algorithm
for numeric input. The class is intended to use with
}
\seealso{
\code{\link{runibic}}
}
