% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\docType{class}
\name{BCUnibicD-class}
\alias{BCUnibicD-class}
\title{Class BCUnibicD}
\description{
An S4 class \code{\link{BCUnibicD-class}} defines UniBic biclustering algorithm
for discrete input.
}
\seealso{
\code{\link{runibic}}
}
