% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\name{BCUnibicD}
\alias{BCUnibicD}
\title{BCUnibicD}
\usage{
BCUnibicD(x, t = 0.95, q = 0, f = 1, nbic = 100, div = 0)
}
\arguments{
\item{x}{a discrete matrix}

\item{t}{consistency level of the block (0.5-1.0]}

\item{q}{a double value for quantile discretization}

\item{f}{filtering overlapping blocks (default 1 do not remove any blocks)}

\item{nbic}{maximum number of biclusters in output}

\item{div}{number of ranks for up(down)-regulated genes: default: 0==ncol(x)}
}
\value{
[biclust::Biclust](https://cran.r-project.org/web/packages/biclust/biclust.pdf) object with detected biclusters
}
\description{
Both \code{BCUnibic} and \code{BCUnibicD} function perform biclustering
using UniBic biclustering algorithm. The major difference 
between both functions is that BCUnibicD requires a discretized matrix, 
whilst BCUnibic could be applied to numeric one.
}
\examples{
A <- runiDiscretize(replicate(10, rnorm(20)))
BCUnibicD(A)
BCUnibicD(A, t = 0.95, q = 0, f = 1, nbic = 100, div = 0)
biclust::biclust(A, method=BCUnibicD(), t = 0.95, q = 0, f = 1, nbic = 100, div = 0)
}
\seealso{
BCUnibic biclustering algrorithm for discrete data
}
