% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateLCS}
\alias{calculateLCS}
\title{This function calculates all pairwise LCSes within the array.}
\usage{
calculateLCS(discreteInput, useFibHeap = TRUE)
}
\arguments{
\item{discreteInput}{is a matrix}

\item{useFibHeap}{boolean value if Fibonacci heap should be used for sorting and seeding}
}
\value{
a list with informa
}
\description{
This function computes unique pairwise Longest Common Subsequences within the matrix.
}
\examples{
calculateLCS(matrix(c(4,3,1,2,5,8,6,7),nrow=2,byrow=TRUE))

}
