% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discretize}
\alias{discretize}
\title{Discretize an input matrix}
\usage{
discretize(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\value{
a discretized matrix containing integers only
}
\description{
This function discretizes the input matrix
}
\examples{
A=replicate(10, rnorm(20))
discretize(A)

}
