% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\docType{package}
\name{runibic}
\alias{runibic}
\alias{runibic-package}
\alias{runibic_d}
\title{runibic: parallel row-based biclustering algorithm
for analysis of gene expression data in R}
\usage{
runibic_d(x, t = 0.95, q = 0, f = 1, nbic = 100, div = 0)

runibic(x, t = 0.95, q = 0, f = 1, nbic = 100, div = 0)
}
\arguments{
\item{x}{numeric matrix or SummarizedExperiment}

\item{t}{consistency level of the block (0.5-1.0].}

\item{q}{a double value for quantile discretization}

\item{f}{filtering overlapping blocks, default 1(do not remove any blocks)}

\item{nbic}{maximum number of biclusters in output}

\item{div}{number of ranks for up(down)-regulated genes: default: 0==ncol(x)}
}
\value{
Biclust object with detected biclusters
}
\description{
runibic: parallel row-based biclustering algorithm
for analysis of gene expression data in R

runibic_d

runibic
}
\section{Functions}{
\itemize{
\item \code{runibic_d}: biclustering algrorithm for discrete data
}}

\examples{
A=matrix(replicate(100, rnorm(100)), nrow=100, byrow=TRUE)
runibic(A)
}
\author{
Patryk Orzechowski, Artur Pańszczyk
}
