% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\name{runibic_d}
\alias{runibic_d}
\title{Parallel row-based biclustering algorithm for analysis of gene expression data in R}
\usage{
runibic_d(x, t = 0.95, q = 0.5, f = 1, nbic = 100, div = 0)
}
\arguments{
\item{x}{integer matrix}

\item{t}{consistency level of the block (0.5-1.0].}

\item{q}{a double value for quantile discretization}

\item{f}{filtering overlapping blocks, default 1(do not remove any blocks)}

\item{nbic}{maximum number of biclusters in output}

\item{div}{number of ranks as which we treat the up(down)-regulated value: default: 0==ncol(x)}
}
\value{
Biclust object with detected biclusters
}
\description{
Parallel row-based biclustering algorithm for analysis of gene expression data in R
}
