% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runibic_params}
\alias{runibic_params}
\title{Set the parameters}
\usage{
runibic_params(t = 0.85, q = 0.5, f = 1, nbic = 100L, div = 0L)
}
\arguments{
\item{t}{consistency level of the block (0.5-1.0]}

\item{q}{a double value for quantile discretization}

\item{f}{filtering overlapping blocks, default 1(do not remove any blocks)}

\item{nbic}{maximum number of biclusters in output}

\item{div}{number of ranks as which we treat the up(down)-regulated value: default: 0==ncol(x)}
}
\value{
NULL (an empty value)
}
\description{
Runibic function for setting parameters
}
\examples{
runibic_params(0.85,100,1,100,0)

}
