% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\name{runibic_se}
\alias{runibic_se}
\title{Parallel row-based biclustering algorithm for analysis of gene expression data in R}
\usage{
runibic_se(x, t = 0.95, q = 0.5, f = 1, nbic = 100, div = 0)
}
\arguments{
\item{x}{SummarizedExperiment}

\item{t}{consistency level of the block (0.5-1.0].}

\item{q}{a double value for quantile discretization}

\item{f}{filtering overlapping blocks, default 1(do not remove any blocks)}

\item{nbic}{maximum number of biclusters in output}

\item{div}{number of ranks as which we treat the up(down)-regulated value: default: 0==ncol(x)}
}
\value{
List of Biclust objects with detected bicluster for each result from every assay within input
}
\description{
Parallel row-based biclustering algorithm for analysis of gene expression data in R
}
\examples{
A=matrix(replicate(100, rnorm(100)), nrow=100, byrow=TRUE)
se = SummarizedExperiment(assays=list(counts=A))
runibic_se(se)
}
