% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\name{BCUnibicD}
\alias{BCUnibicD}
\alias{UnibicD}
\alias{biclust,matrix,BCUnibicD-method}
\alias{BCUnibic}
\alias{Unibic}
\alias{biclust,matrix,BCUnibic-method}
\title{BCUnibicD}
\usage{
\S4method{biclust}{matrix,BCUnibicD}(x=NULL, method=BCUnibicD(), t = 0.95, q = 0.5, f = 1, nbic = 100, div = 0)

\S4method{biclust}{matrix,BCUnibic}(x=NULL, method=BCUnibic(), t = 0.95, q = 0.5, f = 1, nbic = 100, div = 0)
}
\arguments{
\item{x}{numeric matrix}

\item{t}{consistency level of the block (0.5-1.0].}

\item{q}{a double value for quantile discretization}

\item{f}{filtering overlapping blocks, default 1(do not remove any blocks)}

\item{nbic}{maximum number of biclusters in output}

\item{div}{number of ranks as which we treat the up(down)-regulated value: default: 0==ncol(x)}

\item{method}{Unibic method for continuous data}
}
\value{
Biclust object with detected biclusters
}
\description{
\code{BCUnibicD} performs Unibic for a discret matrix.
}
\section{Functions}{
\itemize{
\item \code{BCUnibicD}: biclustering algrorithm for discrete data

\item \code{BCUnibic}: biclustering algrorithm
}}

\examples{
data(BicatYeast)
disc<-discretize(BicatYeast[1:20,1:20])
data(BicatYeast)
}
