% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{backtrackLCS}
\alias{backtrackLCS}
\title{Retrieving a Longest Common Subsequence between two integer vectors.}
\usage{
backtrackLCS(x, y)
}
\arguments{
\item{x}{an integer vector}

\item{y}{an integer vector}
}
\value{
an integer vector containing the the Longest Common Subsequence (LCS)
between vectors x and y (i.e. the values that appear in both x and y in the same order)
}
\description{
This function retrieves the Longest Common Subsequence (LCS)
between two integer vectors by backtracking the matrix 
obtained with dynamic programming.
}
\examples{
A <- c(1, 2, 3, 4, 5)
B <- c(1, 2, 4)
backtrackLCS(A, B)
}
\seealso{
\code{\link{runibic}} \code{\link{pairwiseLCS}} \code{\link{calculateLCS}}
}
