% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster}
\alias{cluster}
\title{Calculating biclusters from sorted list of LCS scores}
\usage{
cluster(discreteInput, discreteInputValues, scores, geneOne, geneTwo, rowNumber,
  colNumber)
}
\arguments{
\item{discreteInput}{an integer matrix with indices of sorted columns}

\item{discreteInputValues}{an integer matrix with values after discretization}

\item{scores}{a numeric vector}

\item{geneOne}{a numeric vector}

\item{geneTwo}{a numeric vector}

\item{rowNumber}{a int with number of rows}

\item{colNumber}{a int with number of columns}
}
\value{
a number of found clusters
}
\description{
Calculating biclusters from sorted list of LCS scores
}
\examples{
cluster(matrix(c(0,3,1,2,3,1,2,0,1,3,2,0),nrow=4,byrow=TRUE),matrix(c(4,3,1,2,5,8,6,7,9,10,11,12),nrow=4,byrow=TRUE),c(13,12,11,7,5,3),c(0,1,2,0,0,1), c(3,2,3,2,1,3),4,3)

}
