% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runibic.R
\docType{package}
\name{runibic}
\alias{runibic}
\alias{runibic-package}
\title{Parallel row-based biclustering algorithm for analysis of gene expression data in R}
\usage{
runibic(x, t = 0.95, q = 0.5, f = 1, nbic = 100, div = 0)
}
\arguments{
\item{x}{numeric matrix}

\item{t}{consistency level of the block (0.5-1.0].}

\item{q}{a double value for quantile discretization}

\item{f}{filtering overlapping blocks, default 1(do not remove any blocks)}

\item{nbic}{maximum number of biclusters in output}

\item{div}{number of ranks as which we treat the up(down)-regulated value: default: 0==ncol(x)}
}
\value{
Biclust object with detected biclusters
}
\description{
Parallel row-based biclustering algorithm for analysis of gene expression data in R

Parallel row-based biclustering algorithm for analysis of gene expression data in R
}
\examples{
A=matrix(replicate(10, rnorm(20)), nrow=10, byrow=TRUE)
runibic(A)
}
\author{
Patryk Orzechowski
}
